Adjusted the Beat-timing of each Song to hit better. Beatindicator Lines now match BeatChecker again
The animatios of the VibinPlants also fit the Beat-timings

+++ Assets/Scripts/Player/BeatIndicator/HalfCircleIndicatorInitializer.cs
@@ -12,8 +12,8 @@ public class HalfCircleIndicatorInitializer : MonoBehaviour
         for (var index = 0; index < indicators.Length; index++)
         {
             HalfCircleIndicator i = indicators[index];
-            i.start_anim_at_normalized_Time = index* (1.0f / indicators.Length) ;
-            i.animationSpeedMultiplier = 1.0f/indicators.Length;
+            i.start_anim_at_normalized_Time = index* (1.0f / (float)indicators.Length) ;
+            i.animationSpeedMultiplier = 1.0f/ (float)indicators.Length;
         }
     }
 }
